Beginning Register Validation
Began to add register validation to check that the email and password entered match the security standard for how they should be configured.
Made it so all input fields are required so they must be filled out before the form can be completed.
Beginning hashed password - to be continued in next sprint to include further checks