Flask Migrations Fixed
I encountered a weird bug where the migration version of the project was different to the one stored in the database. This broke the tracking and caused me to delete previous migrations and initalize from this point. A txt file has been added to show how to do this reset on your own machine.